dwt mix

+++ mmseg/models/utils/wavelet.py
@@ -12,7 +12,7 @@ def dwt_copy_paste(mask, copy_img, paste_img, alpha=0.5):
         Yl_copy, Yh_copy = xfm(copy_img.detach())
         Yl_paste, Yh_paste = xfm(paste_img.detach())
 
-        mask = F.interpolate(mask.unsqueeze(0), scale_factor=0.5, mode='nearest').squeeze(0)
+        mask = F.interpolate(mask.unsqueeze(0).float(), scale_factor=0.5, mode='nearest').squeeze(0).long()
         Yl_mix = Yl_paste.clone()
         Yl_mix[mask] = Yl_copy[mask] * alpha + Yl_paste[mask] * (1 - alpha)
         Yh_mix = [torch.maximum(a, b) for (a, b) in zip(Yh_copy, Yh_paste)]
